Fluxo de Advisor & Correção Administration
Criado o fluxo para orientador (service e controller) de forma corrigida e atualizada.

Corrigido os principais problemas no pacote de administration. Desconheça-se um método que não foi corrigido.

+++ src/main/java/conselho/estudante/com/projetoconselho/services/administration/CourseService.java
@@ -270,7 +270,11 @@ public class CourseService {
     public CourseResponseDTO editShift(Long id, Long shiftId, User actor) {
         Shift shift = shiftService.getObjectShift(shiftId);
         Course course = repository.findById(id).get();
-        shiftService.removeCourseOfShift(course.getShift().getId(), course.getId(), actor);
+
+        if ( course.getShift() != null ) {
+            shiftService.removeCourseOfShift(course.getShift().getId(), course.getId(), actor);
+        }
+
         shiftService.addCourseToShift(shift.getId(), course.getId(), actor);
         logsService.create(actor, course, Collections.singletonList(new ChangeItem("shift", (Object) course.getShift(), (Object) shift)), "update");
         course.setShift(shift);
